Initial Commit
This coding challenge builds an interpreter that executes commands in a
trivial math language:

  note "this is a comment like command";
  let x = 5;
  let y = 7;
  let z = x + -y;
  print z;

Students must implement MyTokenReader.java which is expected to return
tokens (NumberToken, StringToken, NameToken, and SymbolToken) to the
interpreter for processing.

Through this challenge, the students should gain an appreciation for
how the responsibilities are shared and how the various parts of the
system fit together.